// src/app/core/models/avatar.ts
export interface Avatar {
  uuid?: string;
  name: string;
  title: string;
  health: boolean;
  armor: number;
  hungry: number;
  leader: boolean;
  experience: number;
  money: number;
  statusPoint: number;
  isBanned: boolean;
  bannedReason: string;
  isAdmin: boolean;
  isDead: boolean;
  strength: number;
  endurance: number;
  agility: number;
  accuracy: number;
}